Switch from ? to * in docs and examples
The plan for Dotty is to repurpose `?` for wildcards instead of `_`.
In #91 `*` was added to kind-projecter as alternative placeshold syntax
to `?`, and Dotty now also supports `*` as placeholder syntax for cross
compilation via `-Ykind-projector`.

This PR updates the documentation and examples to use `*` rather than
`?` and adds a warning that the `?` syntax will be deprecated in future.